+++ lib/closure_tree/hierarchy_maintenance.rb
@@ -31,12 +31,12 @@ module ClosureTree
     end
 
     def _ct_before_destroy
-      _ct.with_advisory_lock do
+      #_ct.with_advisory_lock do
         delete_hierarchy_references
         if _ct.options[:dependent] == :nullify
           self.class.find(self.id).children.each { |c| c.rebuild! }
         end
-      end
+      #end
       true # don't prevent destruction
     end
 
@@ -59,7 +59,7 @@ module ClosureTree
     end
 
     def delete_hierarchy_references
-      _ct.with_advisory_lock do
+      #_ct.with_advisory_lock do
         # The crazy double-wrapped sub-subselect works around MySQL's limitation of subselects on the same table that is being mutated.
         # It shouldn't affect performance of postgresql.
         # See http://dev.mysql.com/doc/refman/5.0/en/subquery-errors.html
@@ -74,7 +74,7 @@ module ClosureTree
             ) AS x )
             OR descendant_id = #{_ct.quote(id)}
         SQL
-      end
+      #end
     end
 
     module ClassMethods