+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 4.2.6
+
+* Explicitly added MIT licensing to the gemspec.
+
 ### 4.2.5
 
 * Fix for potential deadlock from ```delete_hierarchy_references``` not being called within an